+++ ChangeLog.md
@@ -1,3 +1,14 @@
+# 0.19.3
+Deprecations:
+* `Query.orderBy()` functions with boolean sort parameter were removed
+
+Bug fixes:
+* NoSuchElementException: List is empty on updating not flushed entities ([#708](https://github.com/JetBrains/Exposed/issues/708))
+* Replace function fails on H2 in MySQL mode when using pool 
+* Delayed DAO update flush causes unique constraint violation ([#717](https://github.com/JetBrains/Exposed/issues/717))
+* `Query.empty()` shouldn't add LIMIT 1 if query is marked as forUpdate() ([#684](https://github.com/JetBrains/Exposed/issues/684))
+
+
 # 0.19.1
 Broken changes:
 * `EntityID`, `IdTable`, `IntIdTable`, `LongIdTable`, `UUIDTable` classes from `exposed-core`